requirements.txt
Flask==3.0.3
gevent==24.2.1
gunicorn==21.2.0
python-dotenv==1.0.1
logging==0.4.9.6
flask_compress==1.14
flask_limiter==3.5.1
flask_cors==4.0.0
flask_talisman==0.7.0
redis==5.0.4